Update our CLI tooling initialization and debug messaging (#367)
* Separate configuration and operations

* Update the debug messaging to use rich panel

* Update util debug message

---------
