+++ biothings/cli/utils.py
@@ -26,24 +26,7 @@ from biothings.utils.workers import upload_worker
 import biothings.utils.inspect as btinspect
 
 
-def get_logger(name=None):
-    """
-    Get a logger with the given name.
-    If name is None, return the root logger.
-    """
-    # basicConfig has been setup in cli.py, so we don't need to do it again here
-    # If everything works as expected, we can delete this block.
-    # logging.basicConfig(
-    #     level="INFO",
-    #     format="%(message)s",
-    #     datefmt="[%X]",
-    #     handlers=[RichHandler(rich_tracebacks=True, show_path=False)],
-    # )
-    logger = logging.getLogger(name)
-    return logger
-
-
-logger = get_logger(name=__name__)
+logger = logging.getLogger(name="biothings-cli")
 
 
 def run_sync_or_async_job(func, *args, **kwargs):
@@ -61,79 +44,6 @@ def run_sync_or_async_job(func, *args, **kwargs):
         return func(*args, **kwargs)
 
 
-def load_plugin_managers(
-    plugin_path: Union[str, pathlib.Path], plugin_name: str = None, data_folder: Union[str, pathlib.Path] = None
-):
-    """
-    Load a data plugin from <plugin_path>, and return a tuple of (dumper_manager, upload_manager)
-    """
-    from biothings import config
-    from biothings.hub.dataload.dumper import DumperManager
-    from biothings.hub.dataload.uploader import UploaderManager
-    from biothings.hub.dataplugin.assistant import LocalAssistant
-    from biothings.hub.dataplugin.manager import DataPluginManager
-    from biothings.utils.hub_db import get_data_plugin
-
-    _plugin_path = pathlib.Path(plugin_path).resolve()
-    config.DATA_PLUGIN_FOLDER = _plugin_path.parent.as_posix()
-    sys.path.append(str(_plugin_path.parent))
-
-    if plugin_name is None:
-        plugin_name = _plugin_path.name
-
-    if data_folder is None:
-        data_folder = pathlib.Path(f"./{plugin_name}")
-    data_folder = pathlib.Path(data_folder).resolve().absolute()
-
-    plugin_manager = DataPluginManager(job_manager=None)
-    dmanager = DumperManager(job_manager=None)
-    upload_manager = UploaderManager(job_manager=None)
-
-    LocalAssistant.data_plugin_manager = plugin_manager
-    LocalAssistant.dumper_manager = dmanager
-    LocalAssistant.uploader_manager = upload_manager
-
-    assistant = LocalAssistant(f"local://{plugin_name}")
-    logger.debug(assistant.plugin_name, plugin_name, _plugin_path.as_posix(), config.DATA_PLUGIN_FOLDER)
-
-    dp = get_data_plugin()
-    dp.remove({"_id": assistant.plugin_name})
-    dp.insert_one(
-        {
-            "_id": assistant.plugin_name,
-            "plugin": {
-                "url": f"local://{plugin_name}",
-                "type": assistant.plugin_type,
-                "active": True,
-            },
-            "download": {
-                "data_folder": str(data_folder),  # tmp path to your data plugin
-            },
-        }
-    )
-    p_loader = assistant.loader
-    p_loader.load_plugin()
-
-    return p_loader.__class__.dumper_manager, assistant.__class__.uploader_manager
-
-
-def get_plugin_name(plugin_name=None, with_working_dir=True):
-    """return a valid plugin name (the folder name contains a data plugin)
-    When plugin_name is provided as None, it use the current working folder.
-    when with_working_dir is True, returns (plugin_name, working_dir) tuple
-    """
-    working_dir = pathlib.Path().resolve()
-    if plugin_name is None:
-        plugin_name = working_dir.name
-    else:
-        valid_names = [f.name for f in os.scandir(working_dir) if f.is_dir() and not f.name.startswith(".")]
-        if not plugin_name or plugin_name not in valid_names:
-            rprint("[red]Please provide your data plugin name! [/red]")
-            rprint("Choose from:\n    " + "\n    ".join(valid_names))
-            raise typer.Exit(code=1)
-    return plugin_name, working_dir if with_working_dir else plugin_name
-
-
 def load_plugin(plugin_name: str = None, dumper: bool = True, uploader: bool = True, logger: logging.Logger = None):
     """
     Return a plugin object for the given plugin_name.
@@ -142,8 +52,6 @@ def load_plugin(plugin_name: str = None, dumper: bool = True, uploader: bool = T
 
     If <plugin_name> is not valid, raise the proper error and exit.
     """
-    logger = logger or get_logger(__name__)
-
     _plugin_name, working_dir = get_plugin_name(plugin_name, with_working_dir=True)
     if plugin_name is None:
         # current working_dir has the data plugin
@@ -186,132 +94,171 @@ def load_plugin(plugin_name: str = None, dumper: bool = True, uploader: bool = T
     return current_plugin
 
 
-########################
-# for create command   #
-########################
+def load_plugin_managers(
+    plugin_path: Union[str, pathlib.Path], plugin_name: str = None, data_folder: Union[str, pathlib.Path] = None
+):
+    """
+    Load a data plugin from <plugin_path>, and return a tuple of (dumper_manager, upload_manager)
+    """
+    from biothings import config
+    from biothings.hub.dataload.dumper import DumperManager
+    from biothings.hub.dataload.uploader import UploaderManager
+    from biothings.hub.dataplugin.assistant import LocalAssistant
+    from biothings.hub.dataplugin.manager import DataPluginManager
+    from biothings.utils.hub_db import get_data_plugin
 
+    _plugin_path = pathlib.Path(plugin_path).resolve()
+    config.DATA_PLUGIN_FOLDER = _plugin_path.parent.as_posix()
+    sys.path.append(str(_plugin_path.parent))
 
-def do_create(name, multi_uploaders=False, parallelizer=False, logger=None):
-    """Create a new data plugin from the template"""
-    logger = logger or get_logger(__name__)
-    working_dir = pathlib.Path().resolve()
-    biothing_source_dir = pathlib.Path(__file__).parent.parent.resolve()
-    template_dir = os.path.join(biothing_source_dir, "hub", "dataplugin", "templates")
-    plugin_dir = os.path.join(working_dir, name)
-    if os.path.isdir(plugin_dir):
-        logger.error("Data plugin with the same name is already exists, please remove it before create")
-        return exit(1)
-    shutil.copytree(template_dir, plugin_dir)
-    # create manifest file
-    loader = tornado.template.Loader(plugin_dir)
-    parsed_template = (
-        loader.load("manifest.yaml.tpl").generate(multi_uploaders=multi_uploaders, parallelizer=parallelizer).decode()
+    if plugin_name is None:
+        plugin_name = _plugin_path.name
+
+    if data_folder is None:
+        data_folder = pathlib.Path(f"./{plugin_name}")
+    data_folder = pathlib.Path(data_folder).resolve().absolute()
+
+    plugin_manager = DataPluginManager(job_manager=None)
+    dmanager = DumperManager(job_manager=None)
+    upload_manager = UploaderManager(job_manager=None)
+
+    LocalAssistant.data_plugin_manager = plugin_manager
+    LocalAssistant.dumper_manager = dmanager
+    LocalAssistant.uploader_manager = upload_manager
+
+    assistant = LocalAssistant(f"local://{plugin_name}")
+    assistant_debug_info = (
+        f"[green]Assistant Plugin Name:[/green][bold] "
+        f"[lightsalmon1]{assistant.plugin_name}[/lightsalmon1]\n"
+        f"[green]Assistant Plugin Path:[/green][bold] "
+        f"[lightsalmon1]{_plugin_path.as_posix()}[/lightsalmon1]\n"
+        f"[green]Data Plugin Folder:[/green][bold] "
+        f"[lightsalmon1]{config.DATA_PLUGIN_FOLDER}[/lightsalmon1]"
     )
-    manifest_file_path = os.path.join(working_dir, name, "manifest.yaml")
-    with open(manifest_file_path, "w") as fh:
-        fh.write(parsed_template)
+    logger.debug(assistant_debug_info, extra={"markup": True})
 
-    # remove manifest template
-    os.unlink(f"{plugin_dir}/manifest.yaml.tpl")
-    if not parallelizer:
-        os.unlink(f"{plugin_dir}/parallelizer.py")
-    logger.info(f"Successfully created data plugin template at: \n {plugin_dir}")
+    dp = get_data_plugin()
+    dp.remove({"_id": assistant.plugin_name})
+    dp.insert_one(
+        {
+            "_id": assistant.plugin_name,
+            "plugin": {
+                "url": f"local://{plugin_name}",
+                "type": assistant.plugin_type,
+                "active": True,
+            },
+            "download": {
+                "data_folder": str(data_folder),  # tmp path to your data plugin
+            },
+        }
+    )
+    p_loader = assistant.loader
+    p_loader.load_plugin()
+
+    return p_loader.__class__.dumper_manager, assistant.__class__.uploader_manager
 
 
-###############################
-# for dump & upload command   #
-###############################
+def get_plugin_name(plugin_name=None, with_working_dir=True):
+    """return a valid plugin name (the folder name contains a data plugin)
+    When plugin_name is provided as None, it use the current working folder.
+    when with_working_dir is True, returns (plugin_name, working_dir) tuple
+    """
+    working_dir = pathlib.Path().resolve()
+    if plugin_name is None:
+        plugin_name = working_dir.name
+    else:
+        valid_names = [f.name for f in os.scandir(working_dir) if f.is_dir() and not f.name.startswith(".")]
+        if not plugin_name or plugin_name not in valid_names:
+            rprint("[red]Please provide your data plugin name! [/red]")
+            rprint("Choose from:\n    " + "\n    ".join(valid_names))
+            raise typer.Exit(code=1)
+    return plugin_name, working_dir if with_working_dir else plugin_name
 
 
-def do_dump(plugin_name=None, show_dumped=True, logger=None):
-    """Perform dump for the given plugin"""
+def show_dumped_files(data_folder: Union[str, pathlib.Path], plugin_name: str) -> None:
+    """
+    A helper function to show the dumped files in the data folder
+    """
+    try:
+        data_folder = pathlib.Path(data_folder).resolve().absolute()
+        list_indent = "\n    - "
+        file_collection_repr = list_indent + list_indent.join(pathobj.name for pathobj in data_folder.iterdir())
+        message = (
+            f"[green]Source:[/green][bold] {plugin_name}[/bold]\n"
+            f"[green]Data Folder:[/green][bold] {data_folder}:[/bold]\n"
+            f"[green]Data Folder Contents:[/green][bold]{file_collection_repr}"
+        )
+    except OSError:
+        message = (
+            f"[green]Source:[/green][bold] {plugin_name}[/bold]\n"
+            f"[green]Data Folder:[/green][bold] {data_folder}:[/bold]\n"
+            "Empty directory"
+        )
+
+    console = Console()
+    console.print(
+        Panel(
+            message,
+            title="[bold]Dump[/bold]",
+            title_align="left",
+        )
+    )
+
+
+def get_uploaded_collections(src_db, uploaders):
+    """
+    A helper function to get the uploaded collections in the source database
+    """
+    uploaded_sources = []
+    archived_sources = []
+    temp_sources = []
+    for item in src_db.collection_names():
+        if item in uploaders:
+            uploaded_sources.append(item)
+        for uploader_name in uploaders:
+            if item.startswith(f"{uploader_name}_archive_"):
+                archived_sources.append(item)
+            if item.startswith(f"{uploader_name}_temp_"):
+                temp_sources.append(item)
+    return uploaded_sources, archived_sources, temp_sources
+
+
+def show_hubdb_content():
+    """
+    Output hubdb content in a pretty format.
+    """
     from biothings import config
     from biothings.utils import hub_db
 
+    console = Console()
     hub_db.setup(config)
-    logger = logger or get_logger(__name__)
-    _plugin = load_plugin(plugin_name, dumper=True, uploader=False, logger=logger)
-    dumper = _plugin.dumper
-    dumper.prepare()
-    run_sync_or_async_job(dumper.create_todump_list, force=True)
-    for item in dumper.to_dump:
-        logger.info('Downloading remote data from "%s"...', item["remote"])
-        dumper.download(item["remote"], item["local"])
-        logger.info('Downloaded locally as "%s"', item["local"])
-    dumper.steps = ["post"]
-    dumper.post_dump()
-    dumper.register_status("success")
-    dumper.release_client()
-    # cleanup
-    # Commented out this line below. we should keep the dump info in src_dump collection for other cmds, e.g. upload, list etc
-    # dumper.src_dump.remove({"_id": dumper.src_name})
-    dp = hub_db.get_data_plugin()
-    dp.remove({"_id": _plugin.plugin_name})
-    data_folder = dumper.new_data_folder
-    if show_dumped:
-        logger.info("[green]Success![/green] :rocket:", extra={"markup": True})
-        show_dumped_files(data_folder, _plugin.plugin_name)
-    return _plugin
+    coll_list = [
+        hub_db.get_data_plugin(),
+        hub_db.get_src_dump(),
+        hub_db.get_src_master(),
+        hub_db.get_hub_config(),
+        hub_db.get_event(),
+    ]
+    hub_db_content = []
 
-
-def do_upload(plugin_name=None, show_uploaded=True, logger=None):
-    """Perform upload for the given list of uploader_classes"""
-
-    logger = logger or get_logger(__name__)
-
-    _plugin = load_plugin(plugin_name, dumper=False, uploader=True, logger=logger)
-    for uploader_cls in _plugin.uploader_classes:
-        uploader = uploader_cls.create(db_conn_info="")
-        uploader.make_temp_collection()
-        uploader.prepare()
-        if not uploader.data_folder or not pathlib.Path(uploader.data_folder).exists():
-            logger.error(
-                'Data folder "%s" for "%s" is empty or does not exist yet. Have you run "dump" yet?',
-                uploader.data_folder,
-                uploader.fullname,
-            )
-            raise typer.Exit(1)
-        else:
-            upload_worker(
-                uploader.fullname,
-                uploader.__class__.storage_class,
-                uploader.load_data,
-                uploader.temp_collection_name,
-                10000,
-                1,
-                uploader.data_folder,
-                db=uploader.db,
-            )
-            uploader.switch_collection()
-            uploader.keep_archive = 3  # keep 3 archived collections, that's probably good enough for CLI, default is 10
-            uploader.clean_archived_collections()
-    if show_uploaded:
-        logger.info("[green]Success![/green] :rocket:", extra={"markup": True})
-        show_uploaded_sources(pathlib.Path(_plugin.data_plugin_dir), _plugin.plugin_name)
-    return _plugin
-
-
-def do_dump_and_upload(plugin_name, logger=None):
-    """Perform both dump and upload for the given plugin"""
-    logger = logger or get_logger(__name__)
-    _plugin = do_dump(plugin_name, show_dumped=False, logger=logger)
-    do_upload(plugin_name, show_uploaded=False, logger=logger)
-    logger.info("[green]Success![/green] :rocket:", extra={"markup": True})
-    show_dumped_files(_plugin.dumper.new_data_folder, _plugin.plugin_name)
-    show_uploaded_sources(pathlib.Path(_plugin.data_plugin_dir), _plugin.plugin_name)
-
-
-########################
-# for inspect command  #
-########################
+    for collection in coll_list:
+        content = collection.find()
+        if content:
+            hub_db_content.append(f"[green]Collection:[/green] [bold]{collection.name}[/bold]\n{pformat(content)}")
+    hub_db_content = "\n".join(hub_db_content)
+    console.print(
+        Panel(
+            hub_db_content,
+            title="[bold]Hubdb[/bold]",
+            title_align="left",
+        )
+    )
 
 
 def process_inspect(source_name, mode, limit, merge, logger, do_validate, output=None):
     """
     Perform inspect for the given source. It's used in do_inspect function below
     """
-    from biothings import config
-    from biothings.utils import hub_db
 
     VALID_INSPECT_MODES = ["jsonschema", "type", "mapping", "stats"]
     mode = mode.split(",")
@@ -480,48 +427,9 @@ def process_inspect(source_name, mode, limit, merge, logger, do_validate, output
         #         current_content = {}
         #     current_content.update(mapping)
         #     fp.write(to_json(current_content, indent=True, sort_keys=True))
-        with open(output, "w") as fp:
+        with open(output, "w", encoding="utf-8") as fp:
             fp.write(to_json(mapping, indent=True, sort_keys=True))
             rprint(f"[green]Successfully wrote the mapping info to the JSON file: [bold]{output}[/bold][/green]")
-
-
-def do_inspect(
-    plugin_name=None, sub_source_name=None, mode="type,stats", limit=None, merge=False, output=None, logger=None
-):
-    """Perform inspection on a data plugin."""
-    logger = logger or get_logger(__name__)
-    if not limit:
-        limit = None
-
-    _plugin = load_plugin(plugin_name, logger=logger)
-    # source_full_name = _plugin.plugin_name if sub_source_name else f"{_plugin.plugin_name}.{sub_source_name}"
-    if len(_plugin.uploader_classes) > 1:
-        if not sub_source_name:
-            logger.error('This is a multiple uploaders data plugin, so "--sub-source-name" must be provided!')
-            logger.error(
-                'Accepted values of "--sub-source-name" are: %s',
-                ", ".join(uploader.name for uploader in _plugin.uploader_classes),
-            )
-            raise typer.Exit(code=1)
-        logger.info(
-            'Inspecting data plugin "%s" (sub_source_name="%s", mode="%s", limit=%s)',
-            _plugin.plugin_name,
-            sub_source_name,
-            mode,
-            limit,
-        )
-    else:
-        logger.info('Inspecting data plugin "%s" (mode="%s", limit=%s)', _plugin.plugin_name, mode, limit)
-    # table_space = get_uploaders(pathlib.Path(f"{working_dir}/{plugin_name}"))
-    table_space = [item.name for item in _plugin.uploader_classes]
-    if sub_source_name and sub_source_name not in table_space:
-        logger.error('Your source name "%s" does not exits', sub_source_name)
-        raise typer.Exit(code=1)
-    if sub_source_name:
-        process_inspect(sub_source_name, mode, limit, merge, logger=logger, do_validate=True, output=output)
-    else:
-        for source_name in table_space:
-            process_inspect(source_name, mode, limit, merge, logger=logger, do_validate=True, output=output)
 
 
 def get_manifest_content(working_dir: Union[str, pathlib.Path]) -> dict:
@@ -544,11 +452,6 @@ def get_manifest_content(working_dir: Union[str, pathlib.Path]) -> dict:
     return manifest
 
 
-########################
-# for list command     #
-########################
-
-
 def get_uploaders(working_dir: pathlib.Path) -> list[str]:
     """
     A helper function to get the uploaders from the manifest file in the working directory
@@ -563,47 +466,6 @@ def get_uploaders(working_dir: pathlib.Path) -> list[str]:
     if uploader_section is None and uploaders_section is not None:
         table_space = [item["name"] for item in uploaders_section]
     return table_space
-
-
-def show_dumped_files(data_folder, plugin_name):
-    """A helper function to show the dumped files in the data folder"""
-    console = Console()
-    if not os.path.isdir(data_folder) or not os.listdir(data_folder):
-        console.print(
-            Panel(
-                f"[green]Source:[/green][bold] {plugin_name}[/bold]\n"
-                + f"[green]Data Folder:[/green][bold] {data_folder}:[/bold]\n"
-                + "Empty file!",
-                title="[bold]Dump[/bold]",
-                title_align="left",
-            )
-        )
-    else:
-        console.print(
-            Panel(
-                f"[green]Source:[/green][bold] {plugin_name}[/bold]\n"
-                + f"[green]Data Folder:[/green][bold] {data_folder}:[/bold]\n    - "
-                + "\n    - ".join(os.listdir(data_folder)),
-                title="[bold]Dump[/bold]",
-                title_align="left",
-            )
-        )
-
-
-def get_uploaded_collections(src_db, uploaders):
-    """A helper function to get the uploaded collections in the source database"""
-    uploaded_sources = []
-    archived_sources = []
-    temp_sources = []
-    for item in src_db.collection_names():
-        if item in uploaders:
-            uploaded_sources.append(item)
-        for uploader_name in uploaders:
-            if item.startswith(f"{uploader_name}_archive_"):
-                archived_sources.append(item)
-            if item.startswith(f"{uploader_name}_temp_"):
-                temp_sources.append(item)
-    return uploaded_sources, archived_sources, temp_sources
 
 
 def show_uploaded_sources(working_dir, plugin_name):
@@ -648,94 +510,23 @@ def show_uploaded_sources(working_dir, plugin_name):
         )
 
 
-def show_hubdb_content():
-    """Output hubdb content in a pretty format."""
-    from biothings import config
-    from biothings.utils import hub_db
-
-    console = Console()
-    hub_db.setup(config)
-    coll_list = [
-        hub_db.get_data_plugin(),
-        hub_db.get_src_dump(),
-        hub_db.get_src_master(),
-        hub_db.get_hub_config(),
-        hub_db.get_event(),
-    ]
-    hub_db_content = []
-
-    for collection in coll_list:
-        content = collection.find()
-        if content:
-            hub_db_content.append(f"[green]Collection:[/green] [bold]{collection.name}[/bold]\n{pformat(content)}")
-    hub_db_content = "\n".join(hub_db_content)
-    console.print(
-        Panel(
-            hub_db_content,
-            title="[bold]Hubdb[/bold]",
-            title_align="left",
-        )
-    )
-
-
-def do_list(plugin_name=None, dump=False, upload=False, hubdb=False, logger=None):
-    """List the dumped files, uploaded sources, or hubdb content."""
-    logger = logger or get_logger(__name__)
-    if dump is False and upload is False and hubdb is False:
-        # if all set to False, we list both dump and upload as the default
-        dump = upload = True
-
-    _plugin = load_plugin(plugin_name, dumper=True, uploader=False, logger=logger)
-    if dump:
-        data_folder = _plugin.dumper.current_data_folder
-        if not data_folder:
-            # data_folder should be saved in hubdb already, if dump has been done successfully first
-            logger.error('Data folder is not available. Please run "dump" first.')
-            # Typically we should not need to use new_data_folder as the data_folder,
-            # but we keep the code commented out below for future reference
-            # utils.run_sync_or_async_job(dumper.create_todump_list, force=True)
-            # data_folder = dumper.new_data_folder
-        else:
-            show_dumped_files(data_folder, _plugin.plugin_name)
-    if upload:
-        show_uploaded_sources(pathlib.Path(_plugin.data_plugin_dir), _plugin.plugin_name)
-    if hubdb:
-        show_hubdb_content()
-
-
-########################
-# for serve command    #
-########################
-
-
 def serve(host, port, plugin_name, table_space):
     """
     Serve a simple API server to query the data plugin source.
     """
-    from .web_app import main
     from biothings import config
     from biothings.utils import hub_db
+    from biothings.cli.web_app import main
 
     src_db = hub_db.get_src_db()
     rprint(f"[green]Serving data plugin source: {plugin_name}[/green]")
     asyncio.run(main(host=host, port=port, db=src_db, table_space=table_space))
 
 
-def do_serve(plugin_name=None, host="localhost", port=9999, logger=None):
-    logger = logger or get_logger(__name__)
-    _plugin = load_plugin(plugin_name, dumper=False, uploader=True, logger=logger)
-    uploader_classes = _plugin.uploader_classes
-    table_space = [item.name for item in uploader_classes]
-    serve(host=host, port=port, plugin_name=_plugin.plugin_name, table_space=table_space)
-
-
-########################
-# for clean command    #
-########################
-
-
 def remove_files_in_folder(folder_path):
-    """Remove all files in a folder."""
+    """
+    Remove all files in a folder.
+    """
     for filename in os.listdir(folder_path):
         file_path = os.path.join(folder_path, filename)
         try:
@@ -746,68 +537,3 @@ def remove_files_in_folder(folder_path):
         except Exception as e:
             rprint("[red]Failed to delete %s. Reason: %s [/red]" % (file_path, e))
     shutil.rmtree(folder_path)
-
-
-def do_clean_dumped_files(data_folder, plugin_name):
-    """Remove all dumped files by a data plugin in the data folder."""
-    if not os.path.isdir(data_folder):
-        rprint(f"[red]Data folder {data_folder} not found! Nothing has been dumped yet[/red]")
-        return
-    if not os.listdir(data_folder):
-        rprint("[red]Empty folder![/red]")
-    else:
-        rprint(f"[green]There are all files dumped by [bold]{plugin_name}[/bold]:[/green]")
-        print("\n".join(os.listdir(data_folder)))
-        delete = typer.confirm("Do you want to delete them?")
-        if not delete:
-            raise typer.Abort()
-        remove_files_in_folder(data_folder)
-        rprint("[green]Deleted![/green]")
-
-
-def do_clean_uploaded_sources(working_dir, plugin_name):
-    """Remove all uploaded sources by a data plugin in the working directory."""
-    from biothings import config
-    from biothings.utils import hub_db
-
-    uploaders = get_uploaders(working_dir)
-    src_db = hub_db.get_src_db()
-    uploaded_sources = []
-    for item in src_db.collection_names():
-        if item in uploaders:
-            uploaded_sources.append(item)
-        for uploader_name in uploaders:
-            if item.startswith(f"{uploader_name}_archive_") or item.startswith(f"{uploader_name}_temp_"):
-                uploaded_sources.append(item)
-    if not uploaded_sources:
-        rprint("[red]No source has been uploaded yet! [/red]")
-    else:
-        rprint(f"[green]There are all sources uploaded by [bold]{plugin_name}[/bold]:[/green]")
-        print("\n".join(uploaded_sources))
-        delete = typer.confirm("Do you want to drop them?")
-        if not delete:
-            raise typer.Abort()
-        for source in uploaded_sources:
-            src_db[source].drop()
-        rprint("[green]All collections are dropped![/green]")
-
-
-def do_clean(plugin_name=None, dump=False, upload=False, clean_all=False, logger=None):
-    """Clean the dumped files, uploaded sources, or both."""
-    logger = logger or get_logger(__name__)
-    if clean_all:
-        dump = upload = True
-    if dump is False and upload is False:
-        logger.error("Please provide at least one of following option: --dump, --upload, --all")
-        raise typer.Exit(1)
-
-    _plugin = load_plugin(plugin_name, dumper=True, uploader=False, logger=logger)
-
-    if dump:
-        data_folder = _plugin.dumper.current_data_folder
-        if not data_folder:
-            # data_folder should be saved in hubdb already, if dump has been done successfully first
-            logger.error('Data folder is not available. Please run "dump" first.')
-        do_clean_dumped_files(data_folder, _plugin.plugin_name)
-    if upload:
-        do_clean_uploaded_sources(_plugin.data_plugin_dir, _plugin.plugin_name)
